sudokuSolver/requirements.txt
asgiref==3.5.1
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
dj-database-url==0.5.0
Django==3.0
django-crispy-forms==1.14.0
django-markdownx==3.0.1
fonttools==4.33.3
idna==3.3
importlib-metadata==4.11.3
kiwisolver==1.4.2
Markdown==3.3.7
matplotlib==3.5.2
numpy==1.21.6
opencv-python==4.5.5.64
packaging==21.3
Pillow==9.1.0
psycopg2==2.9.3
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
torch==1.11.0
torchvision==0.12.0
typing_extensions==4.2.0
urllib3==1.26.9
waitress==2.1.1
whitenoise==6.0.0
zipp==3.8.0
